requirements.txt
flake8==5.0.4
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
